# docs/swagger.yaml
definitions:
...
  models.LoginRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: "123456"
        type: string
    type: object
...
  models.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      fullname:
        example: John Doe
        type: string
      password:
        example: "123456"
        type: string
      role:
        example: customer
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Login success
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Wrong password
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login user
      tags:
      - Authentication
...
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success register
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"